% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2bin.R
\name{int2bin}
\alias{int2bin}
\alias{bin2int}
\title{Binary to integer conversions}
\usage{
int2bin( digits, integers )
}
\arguments{
\item{digits}{An integer specifying the number of bits to return.}

\item{binaries}{A matrix of binary integers to convert to base 10.}

\item{integers}{A vector of base 10 integers to convert to binary.}
}
\value{
\code{int2bin()} returns an \code{n x m} matrix. Columns are ordered as in the \code{integers} argument. Each column is an binary number. Each row is a digit of the binary number with total number of rows given by \code{digits}.

\code{bin2int()} returns a vector of integers. Semantic inverse of \code{int2bin()}.
}
\description{
Convert binary vectors to integers and the reverse.
}
\examples{
\dontrun{
x = c(0, 1023)
y = int2bin( 10, x )
z = bin2int( 10, y )
}

}
