% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ixyi.R
\name{ixyi}
\alias{ixyi}
\alias{inter_event_interval}
\title{Compute the time between events.}
\usage{
ixyi( event, time, x_events, y_events, break_events )
}

\arguments{
\item{time}{A numeric vector of times.}

\item{event}{A character vector of events.}

\item{x_events}{A character vector naming the event that starts the inter-event interval.}

\item{y_events}{A character vector naming the event that terminates the inter-event interval.}

\item{break_labels}{A character vector of events that break the inter-event interval.}
}
\value{
Returns a list containing three vectors:
\describe{
\code{ixyi} contains all the inter-event intervals.
\code{ixyi_type} labels each of the inter-event intervals.
\code{ixyi_counter} Tabulates each of the labels.
}
}
\description{
Compute the time between two events. This is useful for computing things like IRTs (time between the same kind of response), changeover times (times between different kinds of responses), PRPs, etc.
}
\details{
When labelling the inter-event intervals, the order of the categories is as they are entered in the \code{x_events} and \code{y_events} arguments. For example, if \code{x_events = c("respR","respL")} and \code{y_events = c("respR","respL") }, then the labels will be ordered: "RR", "RL", "LR", "LL" for the IRTs. These are one-indexed, so \code{1} would label RR IRTs.

The \code{ixyi_counter} tabulates are in the same order as the labels.
}
\examples{
\dontrun{
# Make an event record
time = 1:10*0.5
events = c( rep( "resp", 3), "rft", rep( "resp", 3), "rft", "resp" )
# Calcualte IRTs without considering intervening reinforcer deliveries
irts1 = ixyi( events, time, "resp", "resp", "" )
# Account for intervening reinforcer deliveries
irts2 = ixyi( events, time, "resp", "resp", "rft" )

# Compute the PRPs
prps = ixyi( events, time, "rft", "resp", "" )

# If we have a concurrent schedule, we have IRTs between LL, LR, RL, and RR responses
events2 = c( "respL", "respR", "respR", "rft", "respR", "respR", "respL", "rft", "respR" )
irts3 = ixyi( events2, time, c("respR","respL"), c("respR","respL"), "rft" )
# Two R-R IRTs, one L-R IRT, and one R-L IRT
}
}
\seealso{
\link{ixyi_table_by_z} for tabulating ixyi's as a function of another variable.
}
