% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ixyi_table_by_z.R
\name{ixyi_table_by_z}
\alias{ixyi_table_by_z}
\title{Tabulate inter-event-intervals on subsets}
\usage{
ixyi_table_by_z( event, x_events, y_events, reset_labels, break_labels
    , bin_labels, max_bins )
}
\arguments{
\item{event}{A numeric vector of events.}

\item{x_events}{A numeric vector naming the event that starts the inter-event interval.}

\item{y_events}{A numeric vector naming the event that terminates the inter-event interval.}

\item{reset_labels}{A numeric vector naming events that reset the binning counter.}

\item{break_labels}{A numeric vector of events that break the inter-event interval.}

\item{bin_labels}{A numeric vector of the events that bin the inter-event interval.}

\item{max_bins}{An integer for the number of bins}
}
\value{
Returns a \code{n x m} matrix, where the \code{n} rows define the bins and the \code{m} columns define combinations of the \code{x_events} and \code{y_events}. For example, if we wanted to tabulate the types of IRTs between two alternatives (L and R) for 10 inter-reinforcer intervals, the columns would be ordered (LL, LR, RL, RR) and the rows would be ordered 1 to 10.
}
\description{
Count the number of ixyi as a fucntion of a different variable. This is useful for computing things like IRTs per inter-reinforcer interval.
}
\details{
ixyi_table_by_z() is provded for numeric arguments. ixyi_table_by_z_() is used when \code{event}, \code{x_events}, \code{y_events}, \code{reset_labels}, \code{break_labels}, \code{bin_labels} are character vectors.
}
