% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_match.R
\name{row_match}
\alias{row_match}
\title{Find matching rows in a match}
\usage{
row_match( mat, ref )
}
\arguments{
\item{mat}{Numeric matrix}

\item{ref}{Numeric vector}
}
\value{
A vector of matched row indices.
}
\description{
For a matrix \code{x}, find the rows that are equal to a vector \code{y}.
}
\details{
\code{row_match} returns all of the rows of \code{mat} that match \code{ref}.

\code{row_match1} returns the first matching row. More efficient if only looking for one match because the function terminates when the first match is found.
}
\examples{
\dontrun{
mat = matrix( 1:9, ncol = 3, byrow = T, nrow = 6 )
ref = 1:3
# Get all matching rows
row_match( mat, ref )
# Get first matching row
row_match1( mat, ref )
}

}
